bug-1812771: move processor rules for java crashes to a new file
This moves java crash rules to a new file because we're going to add
more rules at some point. It also renames JavaCrashRule to
JavaStackTraceRule which is a better name and cleans up the code so it's
less weird.
